move auto-configuration to web.autoconfig package
