+++ opentracing-spring-web-autoconfigure/src/main/java/io/opentracing/contrib/spring/web/autoconfig/TracerAutoConfiguration.java
@@ -1,4 +1,4 @@
-package io.opentracing.contrib.spring.web;
+package io.opentracing.contrib.spring.web.autoconfig;
 
 import java.util.logging.Logger;
