+++ opentracing-spring-web-autoconfigure/src/test/java/io/opentracing/contrib/spring/web/autoconfig/RestTemplateAutoConfigurationTest.java
@@ -1,4 +1,4 @@
-package io.opentracing.contrib.spring.web;
+package io.opentracing.contrib.spring.web.autoconfig;
 
 import org.junit.Assert;
 import org.junit.Test;
